+++ xivo_confd/config.py
@@ -82,6 +82,7 @@ DEFAULT_CONFIG = {
         'agent': True,
         'agent_skill': True,
         'api': True,
+        'application': True,
         'call_log': True,
         'call_filter': True,
         'call_filter_fallback': True,